+++ includes/database.php
@@ -1555,6 +1555,38 @@ function thold_upgrade_database($force = false) {
 			'after'    => 'format_file')
 		);
 
+		db_add_column('thold_template', array(
+			'name'     => 'units_suffix',
+			'type'     => 'varchar(10)',
+			'NULL'     => false,
+			'default'  => '',
+			'after'    => 'show_units')
+		);
+
+		db_add_column('thold_data', array(
+			'name'     => 'units_suffix',
+			'type'     => 'varchar(10)',
+			'NULL'     => false,
+			'default'  => '',
+			'after'    => 'show_units')
+		);
+
+		db_add_column('thold_template', array(
+			'name'     => 'decimals',
+			'type'     => 'int(11)',
+			'NULL'     => false,
+			'default'  => '-1',
+			'after'    => 'units_suffix')
+		);
+
+		db_add_column('thold_data', array(
+			'name'     => 'decimals',
+			'type'     => 'int(11)',
+			'NULL'     => false,
+			'default'  => '-1',
+			'after'    => 'units_suffix')
+		);
+
 		db_execute('UPDATE plugin_notification_lists SET enabled = "on"');
 	}
 
@@ -1632,6 +1664,8 @@ function thold_setup_database() {
 	$data['columns'][] = array('name' => 'syslog_enabled', 'type' => 'char(3)', 'NULL' => false, 'default' => '');
 	$data['columns'][] = array('name' => 'data_type', 'type' => 'int(12)', 'NULL' => false, 'default' => '0');
 	$data['columns'][] = array('name' => 'show_units', 'type' => 'char(3)', 'NULL' => false, 'default' => '');
+	$data['columns'][] = array('name' => 'units_suffix', 'type' => 'varchar(10)', 'NULL' => false, 'default' => '');
+	$data['columns'][] = array('name' => 'decimals', 'type' => 'int(11)', 'NULL' => false, 'default' => '-1');
 	$data['columns'][] = array('name' => 'cdef', 'type' => 'int(11)', 'NULL' => false, 'default' => '0');
 	$data['columns'][] = array('name' => 'percent_ds', 'type' => 'varchar(64)', 'NULL' => false, 'default' => '');
 	$data['columns'][] = array('name' => 'expression', 'type' => 'varchar(512)', 'NULL' => false, 'default' => '');
@@ -1719,6 +1753,8 @@ function thold_setup_database() {
 	$data['columns'][] = array('name' => 'snmp_event_warning_severity', 'type' => 'tinyint(1)', 'NULL' => false, 'default' => '2');
 	$data['columns'][] = array('name' => 'data_type', 'type' => 'int(12)', 'NULL' => false, 'default' => '0');
 	$data['columns'][] = array('name' => 'show_units', 'type' => 'char(3)', 'NULL' => false, 'default' => '');
+	$data['columns'][] = array('name' => 'units_suffix', 'type' => 'varchar(10)', 'NULL' => false, 'default' => '');
+	$data['columns'][] = array('name' => 'decimals', 'type' => 'int(11)', 'NULL' => false, 'default' => '-1');
 	$data['columns'][] = array('name' => 'cdef', 'type' => 'int(11)', 'NULL' => false, 'default' => '0');
 	$data['columns'][] = array('name' => 'percent_ds', 'type' => 'varchar(64)', 'NULL' => false, 'default' => '');
 	$data['columns'][] = array('name' => 'expression', 'type' => 'varchar(512)', 'NULL' => false, 'default' => '');