+++ thold.php
@@ -939,7 +939,9 @@ function list_tholds() {
 				$suffix = true;
 			}
 
-			$show_units = ($thold_data['show_units'] ? true : false);
+			$show_units   = ($thold_data['show_units'] ? true : false);
+			$units_suffix = $thold_data['units_suffix'];
+			$decimals     = $thold_data['decimals'];
 
 			if ($baseu == '') {
 				cacti_log('WARNING: Graph Template for local_graph_id ' . $thold_data['local_graph_id'] . ' has been removed!');
@@ -972,21 +974,21 @@ function list_tholds() {
 			}
 
 			form_selectable_cell($data_source, $thold_data['id'], '', 'right');
-			form_selectable_cell(thold_format_number($thold_data['lastread'], 2, $baseu, $suffix, $show_units), $thold_data['id'], '', 'right');
+			form_selectable_cell(thold_format_number($thold_data['lastread'], $decimals, $baseu, $suffix, $show_units, $units_suffix), $thold_data['id'], '', 'right');
 
 			switch($thold_data['thold_type']) {
 				case 0:
-					form_selectable_cell(thold_format_number($thold_data['thold_warning_hi'], 2, $baseu, $suffix, $show_units) . ' / ' . thold_format_number($thold_data['thold_hi'], 2, $baseu, $suffix, $show_units), $thold_data['id'], '', 'right');
-					form_selectable_cell(thold_format_number($thold_data['thold_warning_low'], 2, $baseu, $suffix, $show_units) . ' / ' . thold_format_number($thold_data['thold_low'], 2, $baseu, $suffix, $show_units), $thold_data['id'], '', 'right');
+					form_selectable_cell(thold_format_number($thold_data['thold_warning_hi'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . ' / ' . thold_format_number($thold_data['thold_hi'], $decimals, $baseu, $suffix, $show_units, $units_suffix), $thold_data['id'], '', 'right');
+					form_selectable_cell(thold_format_number($thold_data['thold_warning_low'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . ' / ' . thold_format_number($thold_data['thold_low'], $decimals, $baseu, $suffix, $show_units, $units_suffix), $thold_data['id'], '', 'right');
 					form_selectable_cell('<i>' . plugin_thold_duration_convert($thold_data['local_data_id'], $thold_data['thold_fail_trigger'], 'alert') . '</i>', $thold_data['id'], '', 'right');
 					form_selectable_cell(__('N/A', 'thold'),  $thold_data['id'], '', 'right');
 
 					break;
 				case 1:
-					$hi = thold_format_number($thold_data['thold_hi'], 2, $baseu, $suffix, $show_units);
-					$hi_var = thold_format_number($thold_data['bl_pct_up'], 2, $baseu, $suffix, $show_units);
-					$low = thold_format_number($thold_data['thold_low'], 2, $baseu, $suffix, $show_units);
-					$low_var = thold_format_number($thold_data['bl_pct_down'], 2, $baseu, $suffix, $show_units);
+					$hi      = thold_format_number($thold_data['thold_hi'], $decimals, $baseu, $suffix, $show_units, $units_suffix);
+					$hi_var  = thold_format_number($thold_data['bl_pct_up'], $decimals, $baseu, $suffix, $show_units, $units_suffix);
+					$low     = thold_format_number($thold_data['thold_low'], $decimals, $baseu, $suffix, $show_units, $units_suffix);
+					$low_var = thold_format_number($thold_data['bl_pct_down'], $decimals, $baseu, $suffix, $show_units, $units_suffix);
 
 					if ($thold_data['bl_type'] == 0) {
 						$suffix = ' %';
@@ -1002,8 +1004,8 @@ function list_tholds() {
 
 					break;
 				case 2:
-					form_selectable_cell(thold_format_number($thold_data['time_warning_hi'], 2, $baseu, $suffix, $show_units) . ' / ' . thold_format_number($thold_data['time_hi'], 2, $baseu, $suffix, $show_units), $thold_data['id'], '', 'right');
-					form_selectable_cell(thold_format_number($thold_data['time_warning_low'], 2, $baseu, $suffix, $show_units) . ' / ' . thold_format_number($thold_data['time_low'], 2, $baseu, $suffix, $show_units), $thold_data['id'], '', 'right');
+					form_selectable_cell(thold_format_number($thold_data['time_warning_hi'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . ' / ' . thold_format_number($thold_data['time_hi'], $decimals, $baseu, $suffix, $show_units, $units_suffix), $thold_data['id'], '', 'right');
+					form_selectable_cell(thold_format_number($thold_data['time_warning_low'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . ' / ' . thold_format_number($thold_data['time_low'], $decimals, $baseu, $suffix, $show_units, $units_suffix), $thold_data['id'], '', 'right');
 					form_selectable_cell('<i>' . __('%d Triggers', $thold_data['time_fail_trigger'], 'thold') . '</i>',  $thold_data['id'], '', 'right');
 					form_selectable_cell('<i>' . plugin_thold_duration_convert($thold_data['local_data_id'], $thold_data['time_fail_length'], 'time') . '</i>', $thold_data['id'], '', 'right');
 
@@ -1243,7 +1245,10 @@ function thold_edit() {
 					} else {
 						$suffix = true;
 					}
-					$show_units = ($td['show_units'] ? true : false);
+
+					$show_units   = ($td['show_units'] ? true : false);
+					$units_suffix = $thold_data['units_suffix'];
+					$decimals     = $thold_data['decimals'];
 
 					if (empty($baseu)) {
 						cacti_log('WARNING: Graph Template for local_graph_id ' . $td['local_graph_id'] . ' has been removed!');
@@ -1281,20 +1286,20 @@ function thold_edit() {
 
 					$cur_setting = '<span style="padding-right:4px;">' . __('Last:', 'thold'). '</span>' .
 						($td['lastread'] == '' ? "<span>" . __('N/A', 'thold') . "</span>":"<span style='color:$color'>" .
-						thold_format_number($td['lastread'], 2, $baseu, $suffix, $show_units) . "</span>");
+						thold_format_number($td['lastread'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . "</span>");
 
 					if ($td['thold_type'] != 1) {
 						if ($td['thold_warning_fail_trigger'] != 0) {
 							if ($td['thold_warning_hi'] != '') {
 								$cur_setting .= '<span style="padding:4px">' . __('WHi:', 'thold') . '</span>' .
 									($td['thold_warning_hi'] == '' ? "<span>" . __('N/A', 'thold') . "</span>" : "<span style='color:darkorange'>" .
-									thold_format_number($td['thold_warning_hi'], 2, $baseu, $suffix, $show_units) . '</span>');
+									thold_format_number($td['thold_warning_hi'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . '</span>');
 							}
 
 							if ($td['thold_warning_low'] != '') {
 								$cur_setting .= '<span style="padding:4px">' . __('WLo:', 'thold') . '</span>' .
 									($td['thold_warning_low'] == '' ? "<span>" . __('N/A', 'thold') . "</span>" : "<span style='color:darkorange'>" .
-									thold_format_number($td['thold_warning_low'], 2, $baseu, $suffix, $show_units) . '</span>');
+									thold_format_number($td['thold_warning_low'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . '</span>');
 							}
 						}
 
@@ -1302,20 +1307,20 @@ function thold_edit() {
 							if ($td['thold_hi'] != '') {
 								$cur_setting .= '<span style="padding:4px">' . __('AHi:', 'thold') . '</span>' .
 									($td['thold_hi'] == '' ? "<span>" . __('N/A', 'thold') . "</span>" : "<span style='color:red'>" .
-									thold_format_number($td['thold_hi'], 2, $baseu, $suffix, $show_units) . '</span>');
+									thold_format_number($td['thold_hi'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . '</span>');
 							}
 
 							if ($td['thold_low'] != '') {
 								$cur_setting .= '<span style="padding:4px">' . __('ALo:', 'thold') . '</span>' .
 									($td['thold_low'] == '' ? "<span>" . __('N/A', 'thold') . "</span>" : "<span style='color:red'>" .
-									thold_format_number($td['thold_low'], 2, $baseu, $suffix, $show_units) . '</span>');
+									thold_format_number($td['thold_low'], $decimals, $baseu, $suffix, $show_units, $units_suffix) . '</span>');
 							}
 						}
 					} else {
-						$hi = thold_format_number($td['thold_hi'], 2, $baseu, $suffix, $show_units);
-						$hi_var = thold_format_number($td['bl_pct_up'], 2, $baseu, $suffix, $show_units);
-						$low = thold_format_number($td['thold_low'], 2, $baseu, $suffix, $show_units);
-						$low_var = thold_format_number($td['bl_pct_down'], 2, $baseu, $suffix, $show_units);
+						$hi      = thold_format_number($td['thold_hi'], $decimals, $baseu, $suffix, $show_units, $units_suffix);
+						$hi_var  = thold_format_number($td['bl_pct_up'], $decimals, $baseu, $suffix, $show_units, $units_suffix);
+						$low     = thold_format_number($td['thold_low'], $decimals, $baseu, $suffix, $show_units, $units_suffix);
+						$low_var = thold_format_number($td['bl_pct_down'], $decimals, $baseu, $suffix, $show_units, $units_suffix);
 
 						if ($td['bl_type'] == 0 || $td['bl_type'] == 2) {
 							$suffix = ' %';
@@ -1379,11 +1384,13 @@ function thold_edit() {
 		$suffix = true;
 	}
 
-	$show_units = (isset($thold_data['show_units']) && $thold_data['show_units'] ? true : false);
+	$show_units   = (isset($thold_data['show_units']) && $thold_data['show_units'] ? true : false);
+	$units_suffix = $thold_data['units_suffix'];
+	$decimals     = $thold_data['decimals'];
 
 	if (isset($thold_data['lastread'])) {
 		$header_text = __('Data Source Item [ %s ] - Current value: [ %s ]',
-			(isset($template_rrd) ? $template_rrd['data_source_name'] : ''), thold_format_number($thold_data['lastread'], 2, $baseu, $suffix, $show_units), 'thold');
+			(isset($template_rrd) ? $template_rrd['data_source_name'] : ''), thold_format_number($thold_data['lastread'], $decimals, $baseu, $suffix, $show_units, $units_suffix), 'thold');
 	} else {
 		$header_text = __('Data Source Item [ %s ] - Current value: [ %s ]',
 			(isset($template_rrd) ? $template_rrd['data_source_name'] : ''), '-', 'thold');
@@ -1855,6 +1862,23 @@ function thold_edit() {
 			'description' => __('Display units for very large or small numbers.', 'thold'),
 			'value' => !empty($thold_data['show_units']) ? $thold_data['show_units'] : '',
 		),
+		'units_suffix' => array(
+			'friendly_name' => __('Force Units Suffix', 'thold'),
+			'method' => 'textbox',
+			'default' => '',
+			'max_length' => 10,
+			'size' => 10,
+			'description' => __('If you wish to display a forced unit suffix on the various Threshold pages and Notifications, enter it here.  The size is limited to no more than 10 characters.', 'thold'),
+			'value' => isset($thold_data['units_suffix']) ? $thold_data['units_suffix'] : ''
+		),
+		'decimals' => array(
+			'friendly_name' => __('Decimal Digits to Display', 'thold'),
+			'method' => 'drop_array',
+			'default' => 'NULL',
+			'description' => __('The number of decimal digits to display for Threshold and Current Values.  The default is auto or 2.', 'thold'),
+			'value' => isset($thold_data['decimals']) ? $thold_data['decimals'] : 0,
+			'array' => $thold_decimal_digits,
+		),
 		'expression' => array(
 			'friendly_name' => __('RPN Expression', 'thold'),
 			'method' => 'textarea',
@@ -2221,6 +2245,7 @@ function thold_edit() {
 		$('#expression').prop('disabled', status);
 		$('#upper_ds').prop('disabled', status);
 		$('#show_units').prop('disabled', status);
+		$('#units_suffix').prop('disabled', status);
 
 		// Email Body options
 		$('#email_body').prop('disabled', status);